Expose :runpath, :imported_libs, :libs, and :musl_sysroot*
... for --musl_kbuild.

Implementation notes: for sysroot, we can't use a generic `:sysroot`
label that select()'s based on target platform, because the cc toolchain
always applies the toolchain transition on all_files, causing the
target platform to be under the exec transition. We may be able to
improve this in the future. For now, just use a specific musl_* name.

Bug: 325514706
Change-Id: Ie253af5d62552782affaa585d7f7374c9e6886ac